Added most of the missing BLAS operations.
